--- fuzzers/aflplusplus_eclipser/description.md
@@ -1,14 +1,0 @@
-# aflplusplus + eclipser 2.0
-
-AFL++ fuzzer instance that uses Eclipser 2.0
-  - PCGUARD instrumentation 
-  - dict2file feature
-  - "fast" power schedule
-  - persistent mode + shared memory test cases
-
-Repository: [https://github.com/AFLplusplus/AFLplusplus/](https://github.com/AFLplusplus/AFLplusplus/)
-Repository: [https://github.com/SoftSec-KAIST/Eclipser](https://github.com/SoftSec-KAIST/Eclipser)
-
-[builder.Dockerfile](builder.Dockerfile)
-[fuzzer.py](fuzzer.py)
-[runner.Dockerfile](runner.Dockerfile)
