--- fuzzers/aflplusplus_havoc/builder.Dockerfile
@@ -1,35 +1,0 @@
-# Copyright 2020 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-ARG parent_image
-FROM $parent_image
-
-# Install libstdc++ to use llvm_mode.
-RUN apt-get update && \
-    apt-get install -y wget libstdc++-5-dev libtool-bin automake flex bison \
-                       libglib2.0-dev libpixman-1-dev python3-setuptools unzip \
-                       apt-utils apt-transport-https ca-certificates
-
-# Download and compile afl++.
-RUN git clone https://github.com/AFLplusplus/AFLplusplus.git /afl && \
-    cd /afl && \
-    git checkout 4a0e0270adafbc583d491dfad74d9378a4c06bf7
-
-# Build without Python support as we don't need it.
-# Set AFL_NO_X86 to skip flaky tests.
-RUN cd /afl && unset CFLAGS && unset CXXFLAGS && \
-    export CC=clang && export AFL_NO_X86=1 && \
-    PYTHON_INCLUDE=/ make INTROSPECTION=1 && make install && \
-    make -C utils/aflpp_driver && \
-    cp utils/aflpp_driver/libAFLDriver.a /
