+++ service/experiment-requests.yaml
@@ -19,6 +19,19 @@
 # You can run "make presubmit" to do basic validation on this file.
 # Please add new experiment requests towards the top of this file.
 
+- experiment: 2021-04-17-aflpp
+  description: "afl++ havoc + qemu mode variants"
+  fuzzers:
+    - honggfuzz
+    - aflplusplus
+    - aflplusplus_optimal
+    - aflplusplus_312
+    - aflplusplus_311
+    - aflplusplus_qemu_cmplog
+    - aflplusplus_qemu
+    - aflplusplus_qemu_inmem
+    - aflplusplus_qemu_cmplog_inmem
+
 - experiment: 2021-04-14-aflpp
   description: "afl++ havoc + qemu mode variants"
   fuzzers:
